project/requirements.txt
requests==2.24.0
gevent==20.6.2
djangoql==0.14.0
django_debug_toolbar==2.2
django_debug_toolbar_force==0.1.8
django_cors_headers==3.4.0
Django==3.2.4
simplejson==3.17.2
djangorestframework==3.11.1
pycryptodome==3.10.1
PyJWT==1.7.1
flake8==3.8.3
gunicorn==20.0.4
python-dotenv==0.17.1
google-auth==1.30.1
